requirements.txt
Django==1.8
django-admin-bootstrapped==2.5.6
django-jet==0.1.3
django-markdown==0.8.4
django-mysql==1.0.1
django-taggit==0.17.3
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
Markdown==2.6.5
mysqlclient==1.3.7
Pillow==3.0.0
six==1.10.0
wheel==0.24.0
